Modularizando y optimizando el codigo
