+++ src/extensions.py
@@ -1,0 +1,5 @@
+from flask_mysqldb import MySQL
+from flask_login import LoginManager
+
+mysql = MySQL()
+login_manager = LoginManager()
